Implement Overwrite option
Add a option to plugin, so that when a lastModifie is explictly set or
marked as modified, the plug will not overwrite that value. Not sure if
"overwrite" is the best name. Implementation with default of true will
maintain backward compatibility.

Added a couple tests around the new "overwrite" option.

Also, linting/beautification.
Switched out `underscore` for `lodash`.
And I love 'precommit-hook' for linting and test enforcement, so I threw
that in as well.